detect colours
it finds the overlap between each square and each colour mask. it takes the matrix norm of each one and then assumes the colour in the square is the one with the largest martix norm. due to so many masks the camera has become laggy which is proving quite problematic