removed camera lag
in order to remove the camera lag i changed the sequence of code. instead of checking for colours every frame, it only does it when space is pressed. this however removes the ability to add a feature where it shows the detected colours before scanning them in and moving onto manual changes. prehaps i can make it so that it scans every 60 frames instead, to give a rough guide of what colours are being detected