scramble generator
created a scramble generator. it generates scrambles whenever the timer section is opened or the user presses the button to generate a new scramble. an unintended feature was that when you press it it looks like you can see the scramble being generated, this is as the function is called several times in one button press (takes longer than 1 frame), however i really like this so i will not alter this