rotating pieces error
rubiks cubes moves are considered of going in positive direction if they rotate clockwise relative to the center of the cube. in order to change this i will need to add code which rotates in a different direction if the axis value is equal to -1. this also means the z values have been labelled incorrectly and i will have to change swap the positive and negative directions for that particular axis